// front/lib/generated/asset_registry/types/vestingSeriesStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum VestingSeriesStatus {
  Active,
  Cancelled,
  Draft,
}

export type VestingSeriesStatusArgs = VestingSeriesStatus;

export function getVestingSeriesStatusEncoder(): FixedSizeEncoder<VestingSeriesStatusArgs> {
  return getEnumEncoder(VestingSeriesStatus);
}

export function getVestingSeriesStatusDecoder(): FixedSizeDecoder<VestingSeriesStatus> {
  return getEnumDecoder(VestingSeriesStatus);
}

export function getVestingSeriesStatusCodec(): FixedSizeCodec<
  VestingSeriesStatusArgs,
  VestingSeriesStatus
> {
  return combineCodec(
    getVestingSeriesStatusEncoder(),
    getVestingSeriesStatusDecoder(),
  );
}